requirements.txt
# requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.3.0
opencv-python==4.8.1.78
moviepy==1.0.3
pillow==10.1.0
numpy==1.24.3
requests==2.31.0
aiohttp==3.9.1
schedule==1.2.0
redis==5.0.1
python-dotenv==1.0.0
cryptography==41.0.8
sqlalchemy==2.0.23
alembic==1.13.1
celery==5.3.4
gunicorn==21.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
Flask-CORS==4.0.0
google-api-python-client==2.100.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.1
google-auth==2.23.0
Werkzeug==2.3.7

